+++ PDQ_ILI9341/PDQ_ILI9341.h
@@ -162,7 +162,7 @@ class PDQ_ILI9341 : public PDQ_GFX<PDQ_ILI9341>
 	static void inline begin(void);
 	static void setAddrWindow(uint16_t x0, uint16_t y0, uint16_t x1, uint16_t y1);
 	static void pushColor(uint16_t color);
-	static void pushColor(uint16_t color, int16_t cnt);
+	static void pushColor(uint16_t color, int cnt);
 
 	// Pass 8-bit (each) R,G,B, get back 16-bit packed color
 	static INLINE uint16_t color565(uint8_t r, uint8_t g, uint8_t b)
@@ -175,9 +175,9 @@ class PDQ_ILI9341 : public PDQ_GFX<PDQ_ILI9341>
 	}
 	
 	// required driver primitive methods (all except drawPixel can call generic version in PDQ_GFX with "_" postfix).
-	static void drawPixel(int16_t x, int16_t y, uint16_t color);
-	static void drawFastVLine(int16_t x, int16_t y, int16_t h, uint16_t color);
-	static void drawFastHLine(int16_t x, int16_t y, int16_t w, uint16_t color);
+	static void drawPixel(int x, int y, uint16_t color);
+	static void drawFastVLine(int x, int y, int h, uint16_t color);
+	static void drawFastHLine(int x, int y, int w, uint16_t color);
 	static void setRotation(uint8_t r);
 	static void invertDisplay(boolean i);
 
@@ -186,8 +186,8 @@ class PDQ_ILI9341 : public PDQ_GFX<PDQ_ILI9341>
 		fillScreen_(color);			// call generic version
 	}
 
-	static void drawLine(int16_t x0, int16_t y0, int16_t x1, int16_t y1, uint16_t color);
-	static void fillRect(int16_t x, int16_t y, int16_t w, int16_t h, uint16_t color);
+	static void drawLine(int x0, int y0, int x1, int y1, uint16_t color);
+	static void fillRect(int x, int y, int w, int h, uint16_t color);
 
 	// === lower-level internal routines =========
 	static void commandList(const uint8_t *addr);
@@ -358,7 +358,7 @@ class PDQ_ILI9341 : public PDQ_GFX<PDQ_ILI9341>
 	}
 
 	// normal SPI write with minimal hand-tuned delay (assuming max DIV2 SPI rate)
-	static INLINE void spiWrite16(uint16_t data, int16_t count) INLINE_OPT
+	static INLINE void spiWrite16(uint16_t data, int count) INLINE_OPT
 	{
 		uint8_t temp;
 		__asm__ __volatile__
@@ -493,7 +493,7 @@ class PDQ_ILI9341 : public PDQ_GFX<PDQ_ILI9341>
 	{
 		spiWrite16(data);
 	}
-	static INLINE void spiWrite16(uint16_t data, int16_t count) INLINE_OPT
+	static INLINE void spiWrite16(uint16_t data, int count) INLINE_OPT
 	{
 		while (count-- > 0)
 			spiWrite16(data);
@@ -707,7 +707,7 @@ void PDQ_ILI9341::pushColor(uint16_t color)
 	spi_end();
 }
 
-void PDQ_ILI9341::pushColor(uint16_t color, int16_t count)
+void PDQ_ILI9341::pushColor(uint16_t color, int count)
 {
 	spi_begin();
 
@@ -716,7 +716,7 @@ void PDQ_ILI9341::pushColor(uint16_t color, int16_t count)
 	spi_end();
 }
 
-void PDQ_ILI9341::drawPixel(int16_t x, int16_t y, uint16_t color)
+void PDQ_ILI9341::drawPixel(int x, int y, uint16_t color)
 {
 	if ((x < 0) ||(x >= _width) || (y < 0) || (y >= _height))
 		return;
@@ -730,7 +730,7 @@ void PDQ_ILI9341::drawPixel(int16_t x, int16_t y, uint16_t color)
 	spi_end();
 }
 
-void PDQ_ILI9341::drawFastVLine(int16_t x, int16_t y, int16_t h, uint16_t color)
+void PDQ_ILI9341::drawFastVLine(int x, int y, int h, uint16_t color)
 {
 	// clipping
 	if ((x < 0) || (x >= _width) || (y >= _height))
@@ -759,7 +759,7 @@ void PDQ_ILI9341::drawFastVLine(int16_t x, int16_t y, int16_t h, uint16_t color)
 }
 
 
-void PDQ_ILI9341::drawFastHLine(int16_t x, int16_t y, int16_t w, uint16_t color)
+void PDQ_ILI9341::drawFastHLine(int x, int y, int w, uint16_t color)
 {
 	// clipping
 	if ((x >= _width) || (y < 0) || (y >= _height))
@@ -787,7 +787,7 @@ void PDQ_ILI9341::drawFastHLine(int16_t x, int16_t y, int16_t w, uint16_t color)
 	spi_end();
 }
 
-void PDQ_ILI9341::fillRect(int16_t x, int16_t y, int16_t w, int16_t h, uint16_t color)
+void PDQ_ILI9341::fillRect(int x, int y, int w, int h, uint16_t color)
 {
 	// rudimentary clipping (drawChar w/big text requires this)
 	if ((x >= _width) || (y >= _height))
@@ -820,7 +820,7 @@ void PDQ_ILI9341::fillRect(int16_t x, int16_t y, int16_t w, int16_t h, uint16_t 
 }
 
 // Bresenham's algorithm - thx Wikipedia
-void PDQ_ILI9341::drawLine(int16_t x0, int16_t y0, int16_t x1, int16_t y1, uint16_t color)
+void PDQ_ILI9341::drawLine(int x0, int y0, int x1, int y1, uint16_t color)
 {
 #if 0 && defined(__AVR_ATtiny85__) || defined(__AVR_ATtiny45__)
 	drawLine_(x0, y0, x1, y1, color);
@@ -841,11 +841,11 @@ void PDQ_ILI9341::drawLine(int16_t x0, int16_t y0, int16_t x1, int16_t y1, uint1
 	if (x1 < 0)
 		return;
 
-	int16_t dx, dy;
+	int dx, dy;
 	dx = x1 - x0;
 	dy = abs(y1 - y0);
 
-	int16_t err = dx / 2;
+	int err = dx / 2;
 	int8_t ystep;
 
 	if (y0 < y1)
@@ -860,7 +860,7 @@ void PDQ_ILI9341::drawLine(int16_t x0, int16_t y0, int16_t x1, int16_t y1, uint1
 	uint8_t setaddr = 1;
 
 #if 0 && defined(__AVR_ATtiny85__) && !defined(__AVR_ATtiny45__)
-	int16_t	end = steep ? _height-1 : _width-1;
+	int	end = steep ? _height-1 : _width-1;
 	if (x1 > end)
 		x1 = end;
